# A small performance trick